economy with mongoDB/requirements.txt
python-dotenv>=0.21.0
pymongo>=4.3.3
